requirements.txt
spacy
pdfplumber
python-docx
openai 
fuzzywuzzy 
bs4 
requests 
scikit-learn 
google-search-results 
streamlit 
langchain 
speechrecognition 
langchain-groq 
pyaudio 
coqui-tts
gtts 
playsound 
pydub 
openai-whisper 
wave 
langchain-core 
sounddevice 
soundfile 
librosa